+++ sms_sponsorship/readme/HISTORY.rst
@@ -1,3 +1,8 @@
+(2018-10-11)
+~~~~~~~~~~~~
+
+* Add demand planning computation for SMS events
+
 10.0.1.0.0 (2018-07-09)
 ~~~~~~~~~~~~~~~~~~~~~~~